requirements.txt
opencv-python
torch
scikit-image
pillow
torchvision
gTTS
pygame
